+++ core/mm/sub.mk
@@ -3,3 +3,4 @@ srcs-y += fobj.c
 cflags-fobj.c-$(CFG_CORE_PAGE_TAG_AND_IV) := -Wno-missing-noreturn
 srcs-y += file.c
 srcs-y += vm.c
+srcs-y += core_mmu.c
